+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### `Removed` for now removed features
 
 ### `Fixed` for any bugfixes
-- BackyardBriefApiClient typehinted and method parameters checked for validity
+- BackyardBriefApiClient, BackyardGeo, BackyardJson typehinted and method parameters checked for validity
+- BackyardMysqli::queryArray conditions changed to test for `instanceof \mysqli_result`
 
 ### `Security` in case of vulnerabilities